collector/requirements.txt
# Network Discovery & Agentless Monitoring Requirements
# Install: pip install -r requirements.txt

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================

# Network Scanning
python-nmap>=0.7.1
scapy>=2.5.0  # Alternative for ARP scanning

# Windows Monitoring (WMI)
# Note: Windows-only, install conditionally
# pip install WMI; python -c "import wmi"

# Linux Monitoring (SSH)
paramiko>=3.0.0

# SNMP Monitoring
pysnmp>=4.4.12

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Scheduling (for automated collection)
apscheduler>=3.10.0

# Logging
python-json-logger>=2.0.0

# Web Interface (optional)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# ============================================================================
# NEW - PERFORMANCE & SCALABILITY ENHANCEMENTS
# ============================================================================

# Message Queue (RabbitMQ client)
pika>=1.3.2

# Async SSH (10-100x faster than paramiko for concurrent operations)
asyncssh>=2.14.1

# Async Database Driver (already used in API)
asyncpg>=0.29.0

# Monitoring & Observability
prometheus-client>=0.19.0  # Prometheus metrics exporter
structlog>=23.2.0          # Structured logging

# Connection Management
gevent>=23.9.1             # Greenlets for connection pooling

# Task Queue (Alternative/Advanced - Optional)
# celery>=5.3.4            # Uncomment if using Celery instead of RabbitMQ directly
# redis>=5.0.0             # Uncomment if using Celery with Redis backend
